requirements.txt
absl_py==0.13.0
alphafold==2.0.0
Bio==1.5.2
dm_haiku==0.0.4
dm_tree==0.1.6
haiku==0.01
jax==0.2.14
matplotlib==3.2.2
ml_collections==0.1.0
mock==4.0.3
more_itertools==9.0.0
numpy==1.19.5
pandas==1.3.4
protobuf==4.21.11
scipy==1.7.0
session_info==1.0.0
tensorflow==2.9.2
tree==0.2.4
